backend/requirements.txt
# Django
Django>=5.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2>=2.9.9

# Environment configuration
python-decouple>=3.8

# Image handling
Pillow>=10.0.0

# Rich text editor
django-summernote>=0.8.20

# Production
gunicorn>=22.0.0
dj-database-url>=2.3.0
whitenoise>=6.8.0
django-storages[s3]>=1.14.0
